+++ src/jaxsim/api/__init__.py
@@ -4,6 +4,7 @@ from . import (
     actuation_model,
     com,
     contact,
+    contact_model,
     frame,
     integrators,
     joint,
