+++ src/jaxsim/api/ode.py
@@ -46,36 +46,12 @@ def system_acceleration(
         else jnp.zeros((model.number_of_links(), 6))
     ).astype(float)
 
-    # ======================
-    # Compute contact forces
-    # ======================
-
-    W_f_L_terrain = jnp.zeros_like(f_L)
-    contact_state_derivative = {}
-
-    if len(model.kin_dyn_parameters.contact_parameters.body) > 0:
-
-        # Compute the 6D forces W_f ∈ ℝ^{n_L × 6} applied to links due to contact
-        # with the terrain.
-        W_f_L_terrain, contact_state_derivative = js.contact.link_contact_forces(
-            model=model,
-            data=data,
-            link_forces=f_L,
-            joint_torques=joint_torques,
-        )
-
-    W_f_L_total = f_L + W_f_L_terrain
-
-    # Update the contact state data. This is necessary only for the contact models
-    # that require propagation and integration of contact state.
-    contact_state = model.contact_model.update_contact_state(contact_state_derivative)
-
     # Store the link forces in a references object.
     references = js.references.JaxSimModelReferences.build(
         model=model,
         data=data,
         velocity_representation=data.velocity_representation,
-        link_forces=W_f_L_total,
+        link_forces=f_L,
     )
 
     # Compute forward dynamics.
@@ -92,12 +68,13 @@ def system_acceleration(
         link_forces=references.link_forces(model=model, data=data),
     )
 
-    return v̇_WB, s̈, contact_state
+    return v̇_WB, s̈
 
 
 @jax.jit
 @js.common.named_scope
 def system_position_dynamics(
+    model: js.model.JaxSimModel,
     data: js.data.JaxSimModelData,
     baumgarte_quaternion_regularization: jtp.FloatLike = 1.0,
 ) -> tuple[jtp.Vector, jtp.Vector, jtp.Vector]:
@@ -105,6 +82,7 @@ def system_position_dynamics(
     Compute the dynamics of the system position.
 
     Args:
+        model: The model to consider.
         data: The data of the considered model.
         baumgarte_quaternion_regularization:
             The Baumgarte regularization coefficient for adjusting the quaternion norm.
@@ -166,7 +144,7 @@ def system_dynamics(
     """
 
     with data.switch_velocity_representation(velocity_representation=VelRepr.Inertial):
-        W_v̇_WB, s̈, contact_state_derivative = system_acceleration(
+        W_v̇_WB, s̈ = system_acceleration(
             model=model,
             data=data,
             joint_torques=joint_torques,
@@ -174,6 +152,7 @@ def system_dynamics(
         )
 
         W_ṗ_B, W_Q̇_B, ṡ = system_position_dynamics(
+            model=model,
             data=data,
             baumgarte_quaternion_regularization=baumgarte_quaternion_regularization,
         )
@@ -185,5 +164,4 @@ def system_dynamics(
         base_linear_velocity=W_v̇_WB[0:3],
         base_angular_velocity=W_v̇_WB[3:6],
         joint_velocities=s̈,
-        contact_state=contact_state_derivative,
     )
