+++ src/jaxsim/math/__init__.py
@@ -11,4 +11,4 @@ from .joint_model import JointModel, supported_joint_motion  # isort:skip
 
 
 # Define the default standard gravity constant.
-STANDARD_GRAVITY = 9.81
+STANDARD_GRAVITY = -9.81
