+++ src/jaxsim/rbda/__init__.py
@@ -2,7 +2,7 @@ from . import contacts
 from .aba import aba
 from .collidable_points import collidable_points_pos_vel
 from .crba import crba
-from .forward_kinematics import forward_kinematics_model
+from .forward_kinematics import forward_kinematics, forward_kinematics_model
 from .jacobian import (
     jacobian,
     jacobian_derivative_full_doubly_left,
