+++ src/jaxsim/rbda/contacts/__init__.py
@@ -1,9 +1,5 @@
-from . import relaxed_rigid, rigid, soft
+from . import relaxed_rigid
 from .common import ContactModel, ContactsParams
 from .relaxed_rigid import RelaxedRigidContacts, RelaxedRigidContactsParams
-from .rigid import RigidContacts, RigidContactsParams
-from .soft import SoftContacts, SoftContactsParams
 
-ContactParamsTypes = (
-    SoftContactsParams | RigidContactsParams | RelaxedRigidContactsParams
-)
+ContactParamsTypes = RelaxedRigidContactsParams
