+++ src/jaxsim/rbda/utils.py
@@ -132,10 +132,10 @@ def process_inputs(
     if W_Q_B.shape != (4,):
         raise ValueError(W_Q_B.shape, (4,))
 
-    # Check that the quaternion does not contain NaN values.
+    # Check that the quaternion does not contain NaNs.
     exceptions.raise_value_error_if(
         condition=jnp.isnan(W_Q_B).any(),
-        msg="A RBDA received a quaternion that contains NaN values.",
+        msg="A RBDA received a quaternion that contains NaNs.",
     )
 
     # Check that the quaternion is unary since our RBDAs make this assumption in order
